Moved removal guard to its own class
