+++ publishing-sdk/src/test/java/com/ably/tracking/publisher/DuplicateTrackableGuardTest.kt
@@ -1,6 +1,7 @@
 package com.ably.tracking.publisher
 
 import com.ably.tracking.TrackableState
+import com.ably.tracking.publisher.guards.DuplicateTrackableGuard
 import kotlinx.coroutines.flow.MutableStateFlow
 import org.junit.Assert
 import org.junit.Before
